# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  example.CreateCarRequestBody:
    properties:
      carType:
        type: integer
      name:
        type: string
    required:
    - carType
    - name
    type: object
  example.ErrorResp:
    properties:
      msg:
        type: string
    type: object
  model.Example:
    properties:
      CarType:
        type: integer
      ff:
        type: string
      name:
        description: Name
        type: string
      userID:
        type: string
    type: object
host: localhost
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /car:
    get:
      consumes:
      - application/json
      description: List car
      operationId: ListCar
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Example'
            type: array
      summary: ListCar
      tags:
      - ex
    post:
      consumes:
      - application/json
      description: Create Car
      operationId: CreateCar
      parameters:
      - description: Create Car
        in: body
        name: ReqeustBody
        required: true
        schema:
          $ref: '#/definitions/example.CreateCarRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Example'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/example.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/example.ErrorResp'
      summary: CreateCar
      tags:
      - ex
swagger: "2.0"